Improving signal compatibility
It seems that many monitors refuse to work with my trick to add/remove a
few clocks in each frame to synchronize the output to the input.
Therefore I try to tightly couple the output frequency to the input
frequency via a PLL. To make this again work, it is necessary to adjust
the output video timings a bit. I hope this is will be more compliant
then...