Use Atari clock to drive the HDMI
Because the PLL on the D-Video board can not properly lock to the 3.5Mhz
from the atari pixel clock, I am now using the PAL carrier frequency
clock (about  4.4Mhz) instead. While this is still a bit out of spec for
the PLL (5Mhz minimum) this seems to work now.